requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
cssselect==1.0.1
Django==1.11.4
django-dajax==0.9.2
django-dajaxice==0.7
django-filter==1.0.4
django-registration==2.2
django-rest-auth==0.9.1
djangorestframework==3.6.3
idna==2.6
lxml==3.8.0
MySQL-python==1.2.5
olefile==0.44
Pillow==4.2.1
pytz==2017.2
requests==2.18.4
six==1.10.0
urllib3==1.22
